Feature: Register new user Change: - move authenticate method into UserServiceImpl - PasswordEncoder move into new class BeanContext to prevent circular injection - Add regex for email, password, gender, phone number into application.properties - Add back in sql scripts.
