+++ src/main/java/com/cg_vibely_social_service/CgVibelySocialServiceApplication.java
@@ -1,0 +1,13 @@
+package com.cg_vibely_social_service;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication()
+public class CgVibelySocialServiceApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(CgVibelySocialServiceApplication.class, args);
+    }
+
+}
