+++ src/main/java/com/cg_vibely_social_service/entity/Gender.java
@@ -1,0 +1,7 @@
+package com.cg_vibely_social_service.entity;
+
+public enum Gender {
+    MALE,
+    FEMALE,
+    OTHER
+}
