+++ src/main/java/com/cg_vibely_social_service/security/JwtFilter.java
@@ -1,0 +1,68 @@
+package com.cg_vibely_social_service.security;
+
+import com.cg_vibely_social_service.entity.User;
+import com.cg_vibely_social_service.service.UserService;
+import io.jsonwebtoken.ExpiredJwtException;
+import jakarta.servlet.FilterChain;
+import jakarta.servlet.ServletException;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.web.authentication.WebAuthenticationDetailsSource;
+import org.springframework.stereotype.Component;
+import org.springframework.web.filter.OncePerRequestFilter;
+
+import java.io.IOException;
+
+@Component
+public class JwtFilter extends OncePerRequestFilter {
+    @Autowired
+    private JwtUtil jwtUtil;
+
+    @Autowired
+    private UserService userService;
+
+    @Override
+    protected void doFilterInternal(HttpServletRequest request, HttpServletResponse response, FilterChain filterChain) throws ServletException, IOException {
+        String authorizeHeader = request.getHeader("Authorization");
+        try {
+            String email = null;
+            String jwtToken = null;
+
+            if (authorizeHeader != null && authorizeHeader.startsWith("Bearer ")) {
+                jwtToken = authorizeHeader.substring(7);
+                email = jwtUtil.extractEmail(jwtToken);
+
+                System.out.println(jwtToken);
+                if (email != null) {
+                    User user = (User) userService.loadUserByUsername(email);
+                    if (user != null) {
+                        System.out.println(jwtUtil.isTokenValid(jwtToken));
+                        if (jwtUtil.isTokenValid(jwtToken)) {
+                            UsernamePasswordAuthenticationToken authenticationToken =
+                                    new UsernamePasswordAuthenticationToken(user, null, user.getAuthorities());
+
+                            authenticationToken.setDetails(new WebAuthenticationDetailsSource().buildDetails(request));
+                            SecurityContextHolder.getContext().setAuthentication(authenticationToken);
+                        }
+                        else{
+                            response.setStatus(HttpServletResponse.SC_UNAUTHORIZED);
+                            response.getWriter().write("Unauthorized: Authentication failed");
+                        }
+                    }
+                }
+            }
+        } catch (ExpiredJwtException e) {
+//            response.setContentType("application/json");
+//            response.setStatus(HttpServletResponse.SC_UNAUTHORIZED);
+//            response.sendError(401, "Access token expired");
+//            response.setStatus(401);
+            response.setHeader("Unauthorized", "true");
+            response.setStatus(HttpServletResponse.SC_UNAUTHORIZED);
+        }
+
+        filterChain.doFilter(request, response);
+    }
+}
