+++ src/main/java/com/cg_vibely_social_service/service/RoleService.java
@@ -1,0 +1,4 @@
+package com.cg_vibely_social_service.service;
+
+public interface RoleService {
+}
