--- src/main/java/com/demo/entity/Gender.java
@@ -1,7 +1,0 @@
-package com.demo.entity;
-
-public enum Gender {
-    MALE,
-    FEMALE,
-    OTHER
-}
