--- src/main/java/com/demo/repository/UserRepository.java
@@ -1,12 +1,0 @@
-package com.demo.repository;
-
-import com.demo.entity.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.Optional;
-
-@Repository
-public interface UserRepository extends JpaRepository<User, Long> {
-    Optional<User> findByEmail(String email);
-}
