--- src/main/java/com/demo/security/JwtFilter.java
@@ -1,51 +1,0 @@
-package com.demo.security;
-
-import com.demo.entity.User;
-import com.demo.service.UserService;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
-import org.springframework.security.core.context.SecurityContextHolder;
-import org.springframework.security.web.authentication.WebAuthenticationDetailsSource;
-import org.springframework.stereotype.Component;
-import org.springframework.web.filter.OncePerRequestFilter;
-
-import jakarta.servlet.FilterChain;
-import jakarta.servlet.ServletException;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import java.io.IOException;
-
-@Component
-public class JwtFilter extends OncePerRequestFilter {
-    @Autowired
-    private JwtUtil jwtUtil;
-
-    @Autowired
-    private UserService userService;
-
-    @Override
-    protected void doFilterInternal(HttpServletRequest request, HttpServletResponse response, FilterChain filterChain) throws ServletException, IOException {
-        String authorizeHeader = request.getHeader("Authorization");
-
-        String email = null;
-        String jwtToken = null;
-
-        if (authorizeHeader != null && authorizeHeader.startsWith("Bearer ")) {
-            jwtToken = authorizeHeader.substring(7);
-            email = jwtUtil.extractEmail(jwtToken);
-        }
-
-        if (email != null) {
-            User user = (User) userService.loadUserByUsername(email);
-            if (user != null && jwtUtil.isTokenValid(jwtToken)) {
-                UsernamePasswordAuthenticationToken authenticationToken =
-                        new UsernamePasswordAuthenticationToken(user, null, user.getAuthorities());
-
-                authenticationToken.setDetails(new WebAuthenticationDetailsSource().buildDetails(request));
-                SecurityContextHolder.getContext().setAuthentication(authenticationToken);
-            }
-        }
-
-        filterChain.doFilter(request, response);
-    }
-}
