--- src/main/java/com/demo/service/GenericService.java
@@ -1,17 +1,0 @@
-package com.demo.service;
-
-import jakarta.persistence.EntityNotFoundException;
-import java.util.List;
-import java.util.Optional;
-
-public interface GenericService<T> {
-    List<T> findAll();
-
-    Optional<T> findById(Long id);
-
-    T save(T t);
-
-    T update(Long id, T t) throws EntityNotFoundException;
-
-    void remove(Long id) throws EntityNotFoundException;
-}
