--- src/main/java/com/demo/service/RoleService.java
@@ -1,4 +1,0 @@
-package com.demo.service;
-
-public interface RoleService {
-}
