+++ src/main/resources/application-local.properties
@@ -1,0 +1,24 @@
+#Database
+spring.datasource.url=jdbc:mysql://localhost:3306/vibely_social_webapp
+spring.datasource.username=root
+spring.datasource.password=Cvr*317719#
+spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
+#Hibernate
+spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.MySQL8Dialect
+spring.jpa.properties.hibernate.id.new_generator_mappings = true
+spring.jpa.properties.hibernate.show.sql = true
+spring.jpa.properties.hibernate.format_sql = true
+spring.jpa.generate-ddl=false
+spring.jpa.hibernate.ddl-auto=none
+logging.level.org.hibernate.SQL=DEBUG
+logging.level.org.hibernate.type.descriptor.sql.BasicBinder=TRACE
+
+
+## Spring MVC Properties
+spring.mvc.hiddenmethod.filter.enabled=true
+spring.mvc.throw-exception-if-no-handler-found=true
+
+## DataSource Properties
+# Comma separated list of allowed origins
+app.cors.allowedOrigins=http://localhost:3000
+
