+++ src/main/resources/application.properties
@@ -3,7 +3,8 @@ spring.profiles.active=phat
 
 ## App Properties
 app.jwtSecret=dashdsajfgdskfhdsjbriwgfuyewyrio3y783284y32t21321321
-app.jwtExpirationInMs=1800000
+app.jwtExpirationInMs=180000
+app.jwtRefreshTokenLife=604800000
 ##oauth2
 spring.security.oauth2.client.registration.google.client-id=your-client-id
 spring.security.oauth2.client.registration.google.client-secret=your-client-secret
@@ -11,6 +12,15 @@ spring.security.oauth2.client.registration.google.scope=openid,email
 spring.security.oauth2.client.provider.google.issuer-uri=https://accounts.google.com
 ##security
 spring.security.filter.form-login.enabled=false
+##regex
+user.email.regex=^[\\w-]+(\\.[\\w-]+)*@([\\w-]+\\.)+[a-zA-Z]{2,7}$
+user.password.regex=^(?=.*[a-z])(?=.*[A-Z])(?=.*\\d)(?=.*[@#$%^&+=!])[A-Za-z\\d@#$%^&+=!]{8,16}$
+user.firstName.regex=
+user.lastName.regex=
+user.phone.regex=^[0-9]{10}$
+user.gender.regex=^MALE$|^FEMALE$|^OTHER$
+##pattern
+date.pattern=dd/MM/yyyy
 #firebase
 firebase.bucket-name=bucket-name
 firebase.image-url=https://firebasestorage.googleapis.com/v0/b/${firebase.bucket-name}/o/%s?alt=media
