+++ src/main/resources/script_SQL/insert-data.sql
@@ -1,0 +1,3 @@
+USE vibely_social_webapp;
+
+INSERT INTO role (role_name) VALUES ('ROLE_USER');
