+++ src/main/resources/script_SQL/readme.md
@@ -1,0 +1,14 @@
+###Read me first
+# installation sequence:
+1. schemas-and-users
+2. post-like-comment
+3. chat-and-notification
+Final: insert-data
+
+Note:
+1. Do not edit existing code, only use `ALTER TABLE`
+2. Add description before script code
+        example:
+        ##Chung Nguyen: add column to show in user profile
+        ALTER TABLE user
+                ADD COLUMN age int;
